refactors pipelines, adds google cloud bucket storage functionality to betting_lines collection class.

+++ app/db/collections/box_scores.py
@@ -35,5 +35,8 @@ class BoxScores(BaseCollection):
 
         await self.collection.update_one(query, {'$set': kwargs})
 
-    async def delete_box_scores(self, game_ids: list[str]) -> None:
-        await self.collection.delete_many({'game._id': {'$in': game_ids}})
+    async def delete_box_scores(self, game_ids: list[str] = None):
+        if game_ids:
+            return await self.collection.delete_many({'game._id': {'$in': game_ids}})
+
+        await self.collection.delete_many({})
